Replace geocodio congressional data with `member_of_congress` table
Closes #38.

+++ backend/src/main/kotlin/org/climatechangemakers/act/feature/findlegislator/model/LegislatorArea.kt
@@ -5,5 +5,5 @@ import org.climatechangemakers.act.common.model.RepresentedArea
 
 @Serializable data class LegislatorArea(
   val state: RepresentedArea,
-  val districtNumber: Int?,
+  val districtNumber: Short?,
 )
